(refactor) Rolls VisionDWA back into an RGBD Follower
